hw-backend-summer-2022-3-sql_querying/requirements.txt
asyncpg==0.26.0
attrs==22.1.0
greenlet==1.1.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
pytest-asyncio==0.15.1
PyYAML==6.0
SQLAlchemy==1.4.39
toml==0.10.2
